requirements.txt
Flask==2.3.2
mysql-connector-python==8.1.0
bcrypt==4.0.1
PyJWT==2.8.0
Flask-Cors==4.0.0
